+++ CHANGES.rst
@@ -6,6 +6,9 @@ Version 0.2.0
 
 To be released.
 
+- Added :attr:`Entity.type <wikidata.entity.Entity.type>` property and
+  :class:`~wikidata.entity.EntityType` enum class to represent it.
+
 
 Version 0.1.0
 '''''''''''''